Corrigiendo tipos de Salida

+++ src/Estructura/Division.java
@@ -17,7 +17,7 @@ public class Division extends Tiene2Op implements Operacion {
     }
   
     @Override
-    public float Operar() {
+    public double Operar() {
         return this.Op1/this.Op2;
     }
