+++ src/Estructura/Resta.java
@@ -17,7 +17,7 @@ public class Resta extends Tiene2Op implements Operacion {
     }
 
     @Override
-    public float Operar() {
+    public double Operar() {
         return this.Op1-this.Op2;
     }
