+++ src/Estructura/Suma.java
@@ -18,7 +18,7 @@ public class Suma extends Tiene2Op implements Operacion{
 
     
     @Override
-    public float Operar() {
+    public double Operar() {
         return this.Op1+this.Op2;
     }
